requirements.txt
flask
anthropic
python-dotenv
flask-cors
colorama